[refactor] 캘린더 모달 단일화 / 캘린더와 결제목록 모달 중복 컴포넌트 공용화 / 디자인 수정
