+++ src/components/Common/Modal/Detail/MateDetailModal.js
@@ -9,7 +9,7 @@ import defaultProfile from '@/assets/default_profile.jpg';
 import { calAge, genderToKo } from '@/utils/calculators.js';
 import StarRating from '@/utils/StarRating';
 
-const MateDetailModal = ({ modalData, closeModal, handleApply, pagePosition }) => {
+const MateDetailModal = ({ modalData, closeModal, handleApply, page }) => {
   const { handleContentClick } = useModal();
 
   return (
@@ -36,7 +36,7 @@ const MateDetailModal = ({ modalData, closeModal, handleApply, pagePosition }) =
             </p>
             <p>
               📞
-              {(pagePosition === 'FamilyCalendar' ? modalData.mateResume.tel : `${modalData.tel?.slice(0, 7)}****`) ||
+              {(page === 'calendar' ? modalData.mateResume.tel : `${modalData.tel?.slice(0, 7)}****`) ||
                 '전화번호 정보가 없습니다.'}
             </p>
             <p>📧{modalData.email || modalData.mateResume.email || '이메일 정보가 없습니다.'}</p>
@@ -135,7 +135,7 @@ const MateDetailModal = ({ modalData, closeModal, handleApply, pagePosition }) =
           )}
         </div>
         <div className={styles.button_wrapper}>
-          {pagePosition === 'FamilyCalendar' ? (
+          {page === 'calendar' ? (
             ''
           ) : (
             <button type='submit' onClick={() => handleApply(modalData)}>
