+++ src/components/Common/Modal/Modal.module.css
@@ -21,7 +21,7 @@
   max-width: 660px;
   background-color: white;
   color: #333;
-  padding: 2rem 3rem 0rem 3rem;
+  padding: 2rem 3rem 3rem 3rem;
   border-radius: 10px;
   box-shadow: 0px 6px 20px -4px #0006;
   overflow-y: auto;
@@ -107,6 +107,7 @@ label.with_line {
 }
 
 .button_wrapper {
+  margin: 3rem 0 0 0;
   text-align: center;
 }
 
@@ -136,28 +137,8 @@ label.with_line {
   font-weight: bold;
 }
 
-.button_wrapper {
-  margin: 3rem;
-}
-
 .button_wrapper button {
   background-color: var(--deep-brown);
-}
-
-.location {
-  color: white;
-  border-radius: 10px;
-  padding: 2px 10px 3px 10px;
-  font-weight: bold;
-  margin-right: 5px;
-}
-
-.location.home {
-  background-color: teal;
-}
-
-.location.hospital {
-  background-color: cornflowerblue;
 }
 
 .profile_details p > svg {
@@ -169,4 +150,38 @@ label.with_line {
 .profile_details p > span {
   margin: 0 8px;
   color: var(--normal-gray);
+}
+
+.input_with_button {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.input_with_button > button {
+  padding: 0.6rem 2rem;
+}
+
+.home,
+.hospital {
+  color: white;
+  border-radius: 10px;
+  padding: 2px 10px 3px 10px;
+  font-weight: bold;
+  margin-right: 6px;
+}
+
+.home {
+  background-color: teal;
+}
+
+.hospital {
+  background-color: cornflowerblue;
+}
+
+.introduction {
+  background: var(--transparent-beige);
+  padding: 1rem 1.5rem;
+  border-radius: 10px;
+  color: var(--deep-brown);
 }
