--- src/components/Family/FamilyManage/Calendar/FamilyCalendarModal.module.css
@@ -1,33 +1,0 @@
-.input_with_button {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.input_with_button > button {
-  padding: 0.6rem 2rem;
-}
-
-.home,
-.hospital {
-  color: white;
-  border-radius: 10px;
-  padding: 2px 10px 3px 10px;
-  font-weight: bold;
-  margin-right: 2px;
-}
-
-.home {
-  background-color: teal;
-}
-
-.hospital {
-  background-color: cornflowerblue;
-}
-
-.introduction {
-  background: var(--transparent-beige);
-  padding: 1rem 1.5rem;
-  border-radius: 10px;
-  color: var(--deep-brown);
-}
