--- src/components/Mate/MateManage/Calendar/MateCalendarModal.module.css
@@ -1,23 +1,0 @@
-.home,
-.hospital {
-  color: white;
-  border-radius: 10px;
-  padding: 2px 10px 3px 10px;
-  font-weight: bold;
-  margin-right: 2px;
-}
-
-.home {
-  background-color: teal;
-}
-
-.hospital {
-  background-color: cornflowerblue;
-}
-
-.introduction {
-  background: var(--transparent-beige);
-  padding: 1rem 1.5rem;
-  border-radius: 10px;
-  color: var(--deep-brown);
-}
